+++ examples/react/src/views/index.ts
@@ -1,0 +1,5 @@
+export * from "./dialogs";
+export * from "./menus";
+export * from "./lists";
+export * from "./tabs";
+export * from "./radio";