+++ examples/react/src/views/lists.tsx
@@ -1,0 +1,9 @@
+import { View } from "~/components/view";
+
+export function ListsView() {
+  return (
+    <View headline="Lists" supportingText="Lists are continuous, vertical indexes of text and images.">
+
+    </View>
+  );
+}