+++ examples/react/src/views/menus.tsx
@@ -1,0 +1,9 @@
+import { View } from "~/components/view";
+
+export function MenusView() {
+  return (
+    <View headline="Menus" supportingText="Menus display a list of choices on a temporary surface.">
+
+    </View>
+  );
+}