+++ examples/react/src/views/radio.tsx
@@ -1,0 +1,9 @@
+import { View } from "~/components/view";
+
+export function RadioView() {
+  return (
+    <View headline="Radio button" supportingText="Radio buttons let people select one option from a set of options.">
+
+    </View>
+  );
+}